feat(form): add form components
Several form-related components have been added to the project: FormStep: represents an individual
form step. It is made up of one or more FormField components. FormField: represents a form field. It
consists of a label, a hint and an input. A number of style corrections have also been made to
various components.